microservices/payment-service/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# MongoDB
motor==3.3.2
pymongo==4.6.1

# Stripe Payment Processing
stripe==10.12.0

# Authentication/JWT
python-jose[cryptography]==3.3.0
pyjwt==2.10.0

# HTTP Client
httpx==0.25.1
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
python-multipart==0.0.6

# Monitoring
prometheus-client==0.19.0

# Logging
aiofiles==23.2.1
